kitty: disable shell integration
this is a new feature in version 0.24.0 and I immediately had a problem
with it upon upgrading. Namely that the cursor was changed to a bar when
I was at a command prompt outside of tmux. This bothered me very much.
And apparently shell integration also modified my bashrc.

I'd rather not have to deal with this at all than disable these
individual features.